tests/wpt/web-platform-tests/css/css-fonts/parsing/font-language-override-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Module Level 4: parsing font-language-override with invalid values</title>
<link rel="help" href="https://www.w3.org/TR/css-fonts-4/#propdef-font-language-override">
<meta name="assert" content="font-language-override supports only the grammar 'normal | <string>'.">
<meta name="assert" content="font-language-override string has three letters.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value('font-language-override', 'auto');
test_invalid_value('font-language-override', 'normal "ksw"');

test_invalid_value('font-language-override', '"tr"');
test_invalid_value('font-language-override', '"1 %"');
</script>
</body>
</html>
